tests: move shared helpers into support
The suite documentation already distinguishes reusable cross-context support from fakes that belong beside one test, but the shared Discord, voting, and schema builders still lived under tests/helpers.\n\nMove those three modules to tests/support, update every consumer and the PR-183 dispositions, and document that concrete service subclasses are preferred when adapter constructors name concrete services.
