+++ docs/plans/pr-183-review/13-test-tooling-dispositions.md
@@ -22,21 +22,21 @@ CodeQL's two bot comments on `tests/` are out of scope: they are not review comm
 | 3766500676 | `unit/bot/test_components_v2_ui.py` | "bad asseerts tbh" | **Fixed.** `[17, 1, 1, 1]` became named `discord.ComponentType` members, and the bare button count became a label comparison. |
 | 3771060757 | `unit/bot/submission/test_attachments.py` | "actually we need to verify this claim" | **Fixed.** The absolute claim ("no content type at all") was narrowed to the two shapes actually observed in production, which is also exactly what the design needs and what the two parametrised cases pin. |
 | 3771235233 | `integration/schematics/test_worker_pool.py` | "annoying assert" | **Fixed.** The pinned `nucleation-0.10.1` literal is derived from the installed distribution, so the assertion states its real invariant - the worker loaded the engine this environment installed - instead of failing on every bump. |
-| 3775316974 | `integration/voting/.../test_vote_repository.py` | "having SQL in tests is bad" | **Fixed.** 80 lines of hand-written DDL replaced by the model metadata, via `tests/helpers/schema.py`. The account and guild foreign keys the copy had dropped are now real. Read-back `SELECT`s stay: they are the assertion, not a second setup path. |
+| 3775316974 | `integration/voting/.../test_vote_repository.py` | "having SQL in tests is bad" | **Fixed.** 80 lines of hand-written DDL replaced by the model metadata, via `tests/support/schema.py`. The account and guild foreign keys the copy had dropped are now real. Read-back `SELECT`s stay: they are the assertion, not a second setup path. |
 | 3775329634 | `integration/voting/.../test_vote_repository.py` | "such a long test we should have helpers" | **Fixed.** `attach_vote_message` and `seed_generic_poll` extracted; every value the test asserts on stays in the test body. |
 | 3775333078 | `integration/test_alembic_migrations.py` | "revert" | **Retained.** Reverting would delete the only coverage of a downgrade path, which never runs in production and so is the likeliest to be wrong - the plan's stated bar for reverting is not met. The pinned predecessor revision, which was the real maintenance smell, is now expressed relative to the migration under test. |
 | 3775471758 | `unit/bot/test_command_taxonomy.py` | "hacky as fuck, but maybe justified? needs careful argument" | **Already fixed.** `_check_names` is gone; `_commands_of` carries the argument for reading a cog's tree through `__new__`. |
 | 3775474298 | `unit/bot/test_command_taxonomy.py` | "no registries in tests" | **Already fixed.** The tier test was replaced by `test_sensitive_commands_declare_the_intended_permission_nodes`, which reads nodes off the check predicates. |
 | 3779990071 | `architecture/test_discord_components_v2.py` | "hard coded" | **Fixed.** `owners == ["reactions.py"] * 4` became a name-per-file mapping, so a lost listener and a stray one in a cog are different failures. |
 | 3779992042 | `unit/bot/test_reactions.py` | "subclass protocol" | **Fixed.** The test double subclasses `ReactionSubscriber` and marks its methods `@override`. |
-| 3779995719 | `unit/bot/test_reactions.py` | "we probably need helpers" | **Fixed.** `make_reaction_payload` and `make_reaction_bot` moved to `tests/helpers/discord.py`; the payload is a real `RawReactionActionEvent` and the five `arg-type` suppressions collapsed to one cast in the helper. |
+| 3779995719 | `unit/bot/test_reactions.py` | "we probably need helpers" | **Fixed.** `make_reaction_payload` and `make_reaction_bot` live in `tests/support/discord.py`; the payload is a real `RawReactionActionEvent` and the five `arg-type` suppressions collapsed to one cast in the helper. |
 | 3780229351 | `unit/bot/test_log.py` | "useless tests, remove the whole file" | **Fixed**, by the second half of the instruction. The `cog.log` wrapper test is gone; the privacy test now compares the whole `squid.*` field set, which is what actually proves arguments stay out, and the two listeners with real branching are covered. |
 | 3783323740 | `unit/bot/test_app_main.py` | "refactor the bootstrap to not have a forked process… ugliest test ive ever seen" | **Deferred** to [12-runtime-observability.md](12-runtime-observability.md). The test is ugly because the bootstrap forks; rewriting the test without changing that would only move the ugliness. The process model is a runtime decision, not a test cleanup. |
 | 3783391530 | `unit/test_config.py` | "not part of inert" | **Fixed.** A default `sample_ratio` of 1.0 says nothing about whether anything is exported. Split into the two separate claims it was conflating. |
 | 3783545454 | `integration/observability/test_traces.py` | "i dont really see why we are testing third party code" | **Retained.** Four of its five claims are Squid functions - `correlation_id`, `TraceContextFilter`, `inject_trace_context`, `extracted_trace_span` - whose behaviour exists only once composed with the SDK, and whose unit tests can only assert against a mocked tracer. Rationale recorded in the module docstring. |
 | 3783548057 | `unit/api/test_app.py` | "no." | **Fixed by deletion.** `test_create_app_delegates_optional_instrumentation` asserted that one function calls another and would have passed with instrumentation entirely broken. The real behaviour is covered against a live SDK in `test_traces.py`. |
 | 3783550978 | `unit/test_observability.py` | "no" | **Fixed by deletion**, in part. `test_correlation_id_uses_active_trace_id` mocked the private trace lookup and asserted the value came back; deleted. The untraced fallback is kept and strengthened, because `build_error_presentation` shows that id to users. |
-| 3783707542 | `unit/bot/test_errors.py` | "we really need centralized utils for discord mocks" | **Already fixed.** `test_errors.py` uses `tests/helpers/discord.py`; this plan extended that module rather than starting a second one. |
+| 3783707542 | `unit/bot/test_errors.py` | "we really need centralized utils for discord mocks" | **Already fixed.** `test_errors.py` uses `tests/support/discord.py`; the workspace audit kept genuinely shared support there and context-specific fakes beside their tests. |
 | 3784077161 | `unit/test_config.py` | "maybe don't duplicate the string" | **Fixed.** The anchored line (`cursor_secret`) was retired in `3605d011`; the same pattern in the API pepper and idempotency key id now reads from `BASE_ENVIRONMENT`. |
 | 3784078799 | `unit/test_config.py` | "bad test" | **Fixed.** The exact set of required config groups churned with every new required setting while proving only that several fields were named. Now asserts the property that matters - errors aggregate rather than stopping at the first - with a lower bound on the groups. |
 | 3788032262 | `unit/api/test_phase2_reads.py` | "subclass properly" | **Fixed.** Five duck-typed fakes now subclass the services they replace with `@override`. This surfaced one fake modelling a contract the service does not have: `render_content` returns `bytes` and raises on an unknown hash. |
