+++ docs/plans/pr-183-review/README.md
@@ -62,7 +62,7 @@ plan below as historical, not resolvable with `git show`.
 | 6 | [Schematic domain upload safety](06-schematic-domain-upload-safety.md) | In progress | Updated since the last audit: `IngestRequest.uploaded_by_discord_id` is gone, replaced by a provider-neutral `uploaded_by_account_id: int \| None` (landed alongside the Discord de-privileging work). `RenderRequest.background`'s unlabelled RGBA tuple, the resource-pack value object, and `SimulationRequest` field documentation remain open. |
 | 7 | [Nucleation adapter, wire, and worker hardening](07-nucleation-adapter-worker-hardening.md) | In progress | Worker timeout/queue-wait accounting was already fixed pre-audit. The silent `or SchematicFormat.LITEMATIC` fallback is still live; typed JSON decoding is still open. The plan's cited nucleation pin (`0.10.1`) is stale — the project now pins `0.10.14` — so the upstream #7/#8 reproduction needs to run against the current pin before this can close. |
 | 8 | [Schematic rendering and simulation](08-schematic-rendering-simulation.md) | Done | All six steps re-verified in current code: `prepare_render()` answers `FreshRender \| CachedRender \| SkippedRender`, the ambiguous-input refusal lists the coordinates it will accept, and the stderr pumps are owned by the pool's anyio task group. `SimulationResult` is retained with a recorded rationale. |
-| 9 | [Voting redesign](09-voting-redesign.md) | Mostly done | Subplans 1–4 (domain typing, transport-independent polls, the `/poll` UI, session cleanup) re-verified fully landed — the presentation layer was generalized further than planned, into a shared reconciler/renderer. Subplan 5 is genuinely partial: fixture/builder extraction landed (`tests/helpers/voting.py`, `tests/helpers/schema.py`), but the described cross-product test matrix across `VoteKind`/`VoteTarget`/`VoteVisibility`/`VoteChoice`/`VoteRejection` was never started — only scattered single-axis parametrization exists. |
+| 9 | [Voting redesign](09-voting-redesign.md) | Mostly done | Subplans 1–4 (domain typing, transport-independent polls, the `/poll` UI, session cleanup) re-verified fully landed — the presentation layer was generalized further than planned, into a shared reconciler/renderer. Subplan 5 is genuinely partial: fixture/builder extraction landed (`tests/support/voting.py`, `tests/support/schema.py`), but the described cross-product test matrix across `VoteKind`/`VoteTarget`/`VoteVisibility`/`VoteChoice`/`VoteRejection` was never started — only scattered single-axis parametrization exists. |
 | 10 | [Shared reaction routing](10-shared-reaction-routing.md) | Not started | Re-verified unchanged: `squid/bot/reactions.py` still spawns shard workers with bare `asyncio.create_task` and drives shutdown/subscriber dispatch off raw tasks and `asyncio.gather`, even though `BackgroundTaskSupervisor` is already used elsewhere in the bot. No task-ownership, observability, or contract-narrowing work has landed. |
 | 11 | [API, auth, records, and sync](11-api-auth-records-sync.md) | Done | All eight subplans re-verified landed, and all three recorded caveats (pinned 422 reason phrase, `ReconciliationResource`'s hand-mapped `post_kind`, three surviving "principal" external contracts) remain accurate as documented. The later `contract()`/OPERATIONS-table migration work is unrelated follow-on scope, confirmed to belong to [`rest-api.md`](../completed/rest-api.md), not this plan. |
 | 12 | [Runtime and observability](12-runtime-observability.md) | In progress | Subplan 2 (short correlation-id display reference) confirmed done, and it went on to enable error-report storage with a per-invocation correlation ID bound around every Discord command. Subplan 1 (log-transport documentation) is still not started, though an unrelated Docker fix incidentally settled the log-volume-ownership question by switching to a named volume. Subplans 3–8 (telemetry-guard collapse, worker trace-header field, command-span naming, the `squid/api/errors.py` FastAPI revert, the welcome-relay sleep/correlation fix, and the entry-point lifecycle test rewrite) show no code changes at all. |
