+++ tests/support/__init__.py
@@ -1,0 +1,1 @@
+"""Reusable test support shared across architectural boundaries."""
