+++ tests/unit/voting/application/test_vote_service.py
@@ -29,7 +29,7 @@ from squid.voting.domain import (
     VoteVisibility,
 )
 from squid.voting.errors import InvalidVoteConfigurationError
-from tests.helpers.voting import DEFAULT_BUILD_TARGET, GENERIC_OPTIONS, build_snapshot, poll_snapshot
+from tests.support.voting import DEFAULT_BUILD_TARGET, GENERIC_OPTIONS, build_snapshot, poll_snapshot
 
 STAFF = frozenset({VOTE_WEIGHT_STAFF.name})
 DELETE_LOG = frozenset({VOTE_LOG_DELETE_CAST.name})
